+++ src/main/java/spring/backend/m_bl5_g1_su25/OnlineShopping/Cart/service/CartItemService.java
@@ -17,7 +17,7 @@ public interface CartItemService {
     public BigDecimal getCartTotalPrice(Integer userId);
 
     @Transactional
-    public Cart_Items addToCart(Integer userId, Integer productId, Integer quantity);
+    public Cart_Items addToCart(Integer userId, Long productId, Integer quantity);
 
     @Transactional
     public Cart_Items updateCartItem(Integer cartItemId, Integer quantity);
